Do not clone Kokkos from repo in install test
